One last commit
Starting over.  Commiting this last commit so this mess is saved before
starting a new branch.  I go into why on my dev blog which i will add to
the readme.  But basically starting after this i will be attempting a
rewrite based off Teth code.  Identifying abstract functional blocks and
then writing those piece by piece with the code as ref.  It will still
be in C, and maybe still in windows.