ship now properly says when a connection was blocked
However, it appears its not handling disconnects very well.  If the
client is forced closed the server doesnt close the socket, and instead
reads the last message sent.  This may be because of an oddity in how
the dummyclient is written, and not reflect an error in the server.  But
it would be nice to know that the ship can handle a failed connection
(from like a crash) just as well as a proper disconnect (client sending
a dc signal).